Fix broken spec
